move test to right directory
